refactor(cli): extract formatBytes to shared utility and add E2E tests
- Move formatBytes and formatRelativeTime to file-utils.ts
- Update artifact/list.ts, volume/list.ts, and clone-utils.ts to use shared utility
- Add comprehensive E2E tests for list and clone commands
